// Frontend/src/types/centerUser.ts
export interface CenterUserFormData {
  user_name: string;
  user_username: string;
  user_password: string;
  user_email: string;
  center_id: number;
  cu_status: number;
}
export interface CenterData {
  id: number; // Add this field
  center_name: string;
  center_full_name: string;
  center_status: number;
}
export interface CenterData {
  id: number; // Add this field
  center_name: string;
  center_status: number;
}